requirements.txt
openai==1.70.0
requests>=2.28.0
python-dotenv>=0.21.0
